Refactor for clarity and clear concern
