+++ src/main/java/com/concrete/challenge/bean/categories/external/SubCategoryLevel2.java
@@ -1,6 +1,6 @@
 package com.concrete.challenge.bean.categories.external;
 
-import com.concrete.challenge.bean.categories.mobile.behaviour.ConvertToMobile;
+import com.concrete.challenge.bean.categories.structure.behaviour.ConvertToMobile;
 import com.concrete.challenge.bean.categories.structure.AbstractCategory;
 import com.concrete.challenge.bean.categories.structure.Category;
 import com.concrete.challenge.bean.categories.structure.IconImageUrl;
