+++ src/main/java/com/concrete/challenge/bean/categories/structure/behaviour/ConvertToMobile.java
@@ -1,4 +1,4 @@
-package com.concrete.challenge.bean.categories.mobile.behaviour;
+package com.concrete.challenge.bean.categories.structure.behaviour;
 
 import com.concrete.challenge.bean.categories.structure.Category;
 import com.concrete.challenge.converter.CategoryConverter;
